third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/mojo-loading/http/tests/inspector/application-panel/storage-view-reports-quota-expected.txt

CONSOLE MESSAGE: line 124: InspectorTest.IndexedDB_callback1
CONSOLE MESSAGE: line 124: InspectorTest.IndexedDB_callback2
CONSOLE MESSAGE: line 124: InspectorTest.IndexedDB_callback3
Tests quota reporting.

Tree element found: true
Clear storage view is visible: true
0 B storage quota used out of 5.0 MB

Running: Now with data
9.5 MB storage quota used out of 5.0 MB
